RF: Moved spdot and sparse_nnls into dipy.core.optimize
